Move loader registry to target factory
