+++ session.cpp
@@ -30,7 +30,7 @@ Session& Session::configClear()
 	m_targetinfo.clear();
 	m_source_files.clear();
 	m_dirloader.configClear();
-	m_spider.loadersUnregister();
+	m_delegator.loadersUnregister();
 	m_delegator.factoriesUnregister();
 	m_target_hooks.configClear();
 	graphDirtySet();
@@ -255,25 +255,24 @@ namespace
 	class TargetHookRegistrator:public Target_Hook_Registry::EnumCallbackFilenameExt
 		{
 		public:
-			explicit TargetHookRegistrator(Spider& spider,Target_FactoryDelegator& delegator) noexcept:
-				r_spider(spider),r_delegator(delegator)
+			explicit TargetHookRegistrator(Target_FactoryDelegator& delegator) noexcept:
+				r_delegator(delegator)
 				{}
 
 			void operator()(const Stringkey& filename_ext,const Target_Hook& hook)
 				{
-				r_spider.loaderRegister(filename_ext,hook.loaderGet());
-				r_delegator.factoryRegister(filename_ext,hook.factoryGet());
+				r_delegator.loaderRegister(filename_ext,hook.loaderGet())
+					.factoryRegister(filename_ext,hook.factoryGet());
 				}
 		private:
-			Spider& r_spider;
 			Target_FactoryDelegator& r_delegator;
 		};
 	}
 
 void Session::targetHooksRegister() const
 	{
-	m_target_hooks.enumerate(TargetHookRegistrator(m_spider,m_delegator));
-	m_spider.loaderRegister(Stringkey("."),m_dirloader);
+	m_target_hooks.enumerate(TargetHookRegistrator(m_delegator));
+	m_delegator.loaderRegister(Stringkey("."),m_dirloader);
 	targetHooksDirtyClear();
 	}
 
@@ -339,5 +338,5 @@ bool Session::loaderHas(const char* filename) const
 		{return 0;}
 	if(targetHooksDirty())
 		{targetHooksRegister();}
-	return m_spider.loaderHas(Stringkey(filename_ext));
+	return m_delegator.loaderHas(Stringkey(filename_ext));
 	}
