+++ spiderdefault.cpp
@@ -4,8 +4,8 @@
 
 #include "spiderdefault.hpp"
 #include "directorylister.hpp"
-#include "fileinfo.hpp"
 #include "target_loader.hpp"
+#include "target_factorydelegator.hpp"
 
 using namespace Maike;
 
@@ -26,49 +26,6 @@ SpiderDefault& SpiderDefault::scanFile(const char* filename,const char* in_dir)
 	return *this;
 	}
 
-SpiderDefault& SpiderDefault::loaderRegister(const Stringkey& filename_ext
-	,const Target_Loader& loader)
-	{
-	m_r_loaders[filename_ext]=&loader;
-	return *this;
-	}
-
-void SpiderDefault::loadersUnregister() noexcept
-	{
-	m_r_loaders.clear();
-	}
-
-static const Target_Loader* targetLoaderGet(const Stringkey& key
-	,const std::map<Stringkey,const Target_Loader*>& loaders)
-	{
-	auto i=loaders.find(key);
-	if(i==loaders.end())
-		{
-		return nullptr;
-		}
-	return i->second;
-	}
-
-static Stringkey targetLoaderKeyGet(const std::string& filename)
-	{
-	switch(FileInfo(filename.c_str()).typeGet())
-		{
-		case FileInfo::Type::FILE:
-			{
-			auto pos=filename.find_last_of('.');
-			if(pos==std::string::npos)
-				{return Stringkey("");}
-			return Stringkey(&filename[pos]);
-			}
-
-		case FileInfo::Type::DIRECTORY:
-			return Stringkey(".");
-
-		default:
-			return Stringkey("");
-		}
-	}
-
 SpiderDefault& SpiderDefault::run()
 	{
 	while(!m_files_to_scan.empty())
@@ -76,12 +33,8 @@ SpiderDefault& SpiderDefault::run()
 		auto p=std::move(m_files_to_scan.top());
 		m_files_to_scan.pop();
 		auto filename=p.first.c_str();
-		auto loader=targetLoaderGet(targetLoaderKeyGet(p.first),m_r_loaders);
-		if(loader!=nullptr)
-			{
-			auto in_dir=p.second.c_str();
-			loader->targetsLoad(filename,in_dir,*this,r_targets,r_target_creator);
-			}
+		auto in_dir=p.second.c_str();
+		r_target_creator.targetsLoad(filename,in_dir,*this,r_targets);
 		}
 
 	return *this;
