+++ spiderdefault.hpp
@@ -30,14 +30,6 @@ namespace Maike
 			SpiderDefault& scanFile(const char* filename,const char* in_dir);
 			SpiderDefault& run();
 
-			SpiderDefault& loaderRegister(const Stringkey& filename_ext
-				,const Target_Loader& loader);
-
-			void loadersUnregister() noexcept;
-
-			bool loaderHas(const Stringkey& filename_ext) const noexcept
-				{return m_r_loaders.find(filename_ext)!=m_r_loaders.end();}
-
 		private:
 			std::map<Stringkey,const Target_Loader*> m_r_loaders;
 			Target_FactoryDelegator& r_target_creator;
