+++ target_factorydelegator.hpp
@@ -11,10 +11,12 @@ namespace Maike
 	{
 	class Target;
 	class Target_Factory;
+	class Target_Loader;
 	class Stringkey;
 	class ResourceObject;
 	class Dependency;
 	class DependencyGraph;
+	class Spider;
 
 	class PRIVATE Target_FactoryDelegator
 		{
@@ -47,14 +49,23 @@ namespace Maike
 				,DependencyCollector& cb,DependencyGraph& graph)=0;
 
 
+			virtual size_t idGet() noexcept=0;
+
+			virtual const char* rootGet() const noexcept=0;
+
+			virtual void targetsLoad(const char* filename,const char* in_dir,Spider& spider
+				,DependencyGraph& targets)=0;
+
+
 			virtual Target_FactoryDelegator& factoryRegister(const Stringkey& filename_ext
 				,const Target_Factory& factory)=0;
 
 			virtual void factoriesUnregister()=0;
 
-			virtual size_t idGet() noexcept=0;
+			virtual Target_FactoryDelegator& loaderRegister(const Stringkey& filename_ext
+				,const Target_Loader& loader)=0;
 
-			virtual const char* rootGet() const noexcept=0;
+			virtual void loadersUnregister() noexcept=0;
 
 		protected:
 			~Target_FactoryDelegator()=default;
