+++ versioninfo.txt
@@ -1,1 +1,1 @@
-0.13.0-2-gf938931-dirty
+0.13.0-3-g52fd395-dirty
